pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "parrrot"
version = "0.1.0"
description = "Your private Jarvis. Runs on your machine. Does anything."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Parrrot Contributors" }]
keywords = ["ai", "assistant", "local", "ollama", "automation", "jarvis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # CLI + UI
    "typer>=0.12.0",
    "rich>=13.7.0",
    "prompt_toolkit>=3.0.0",
    # CDP browser control (connects to Chrome/Edge — fast web nav, all languages)
    "websocket-client>=1.7.0",
    # Config
    "tomli>=2.0.0; python_version < '3.11'",
    "tomli-w>=1.0.0",
    # HTTP / LLM
    "httpx>=0.27.0",
    "ollama>=0.2.0",
    # File system watching
    "watchdog>=4.0.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Desktop notifications
    "plyer>=2.1.0",
    # PC control
    "pyautogui>=0.9.54",
    "keyboard>=0.13.5",
    "mouse>=0.7.1",
    # Windows UI Automation — reads accessibility tree for reliable PC navigation
    "pywinauto>=0.6.8",
    "comtypes>=1.4.0",
    # Screenshot + OCR
    "Pillow>=10.3.0",
    "mss>=9.0.2",
    # Process management
    "psutil>=5.9.0",
    # Encryption for secrets
    "cryptography>=42.0.0",
    # Fuzzy matching for memory search
    "rapidfuzz>=3.9.0",
    # TOML writing
    "tomli-w>=1.0.0",
    # Date parsing
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.44.0",
]
ocr = [
    "pytesseract>=0.3.13",
]
youtube = [
    "yt-dlp>=2024.5.0",
]
email = [
    # uses stdlib imaplib/smtplib — no extra deps
]
voice = [
    "pyttsx3>=2.90",
    "SpeechRecognition>=3.10.0",
    "pyaudio>=0.2.14",
]
all = [
    "parrrot[browser,ocr,youtube,voice]",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
parrrot = "parrrot.main:app"

[tool.hatch.build.targets.wheel]
packages = ["parrrot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
